Move AccessSelector out of the root components directory.
Move UserType pipe to AccessSelectorModule
